Ensure scratch operations happen on master core.
Reviewed-on: http://gfw160.austin.ibm.com:8080/gerrit/275
Tested-by: Jenkins Server

+++ src/include/kernel/types.h
@@ -25,11 +25,11 @@
 
 #include <stdint.h>
 
-typedef uint16_t 	tid_t;	// This is 16-bit for the VMM mapping of 
+typedef uint16_t 	tid_t;	// This is 16-bit for the VMM mapping of
 				// stacks.  See VmmManager.
 struct task_t;
 
-typedef uint64_t	cpuid_t;
+typedef uint32_t	cpuid_t;
 struct cpu_t;
 
 #endif
